requirement.txt
pandas==3.0.1
dotenv==0.9.9
PyYAML==6.0.3
pyarrow==23.0.1
opencv-python==4.13.0.92
mediapipe==0.10.21